Move version and format checks in a different file
Move the code related to checking the version and the entries of
bunnyfile to a different file than hops/package.go.
